move validator from core to instrument
Since in a lambda function one can't provide a proper docstring I would prefer to keep this as a proper function
